+++ CHANGELOG.md
@@ -11,6 +11,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Added CLI for refreshing app tokens `oauthclient/apps/refresh-tokens <app handle>`
 - Added events for token refresh before, after, and error
 - Added `getValidTokensForUser()` to App model
+- Added `checkTokenWithProvider()` to Credentials service
+- Added `ValidatesToken` interface for providers to implement
 
 ### Fixed
 - Fixed a potential bug getting tokens by app & user